% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_treatment_groups.R
\name{add_treatment_groups}
\alias{add_treatment_groups}
\title{Customize [ dm ] mapping for stats calculation.}
\usage{
add_treatment_groups(mapping, data, stratification_col)
}
\description{
Customize [ dm ] mapping for stats calculation.
}
