% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Get Summary AE Statistics}
\usage{
get_stats(dm, ae, settings, statistic = "Risk Ratio")
}
\arguments{
\item{dm}{`data.frame` Subject-level dataset with one record per subject.}

\item{ae}{`data.frame` Event-level datasets with one record per adverse event.}

\item{settings}{`list` Named list of settings (see examples below for standard list).}

\item{statistic}{`character` Statistic to calculate for AE plot. Options are risk ratio ("RR"
or "Risk Ratio"), risk difference ("RD" or "Risk Difference"). Defaults to
"Risk Ratio".}
}
\value{
a data frame of group-wise statistics for use in the volcano plot
}
\description{
Compares reference and comparison groups to calculate group-wise metrics and p-values for use in AE volcano plot.
}
\examples{
settings <- list(
    stratification_col = "AEBODSYS",
    treatment_col = "ARM",
    reference_group = "Placebo",
    comparison_group = "Xanomeline High Dose",
    id_col = "USUBJID"
)

get_stats(
    dm = safetyData::adam_adsl,
    ae = safetyData::adam_adae,
    settings
)

}
