% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{getStats}
\alias{getStats}
\title{Get statistics for AE data}
\usage{
getStats(dfAE, dfDemog, settings, stat = "RR")
}
\arguments{
\item{settings}{settings}

\item{data}{A data frame with adverse event data and placeholder rows for participants with no AEs}
}
\value{
a data frame for use in the volcano plot
}
\description{
Get statistics for AE data
}
\examples{
settings<-list(
  stratification_col="AEBODSYS",
  group_col="ARM", 
  reference_group="Placebo",
  comparison_group="Xanomeline High Dose",
  id_col="USUBJID"
) 
getStats(dfAE=safetyData::adam_adae, dfDemog = safetyData::adam_adsl, settings)

}
