% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_app.R
\name{volcano_app}
\alias{volcano_app}
\title{Volcano Plot App}
\usage{
volcano_app(
  dm = safetyData::adam_adsl,
  ae = safetyData::adam_adae,
  settings = NULL,
  runNow = TRUE
)
}
\arguments{
\item{dm}{`data.frame` Subject-level data with one row per subject.}

\item{ae}{`data.frame` Event-level data with one row per adverse event.}

\item{settings}{`list` `{safetyGraphics}` settings}

\item{runNow}{`logical` run app immediately?}
}
\value{
Initializes Shiny app. No return value.
}
\description{
Initializes a stand-alone volcano plot shiny application.
}
