% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_server.R
\name{volcano_server}
\alias{volcano_server}
\title{Volcano Plot Module - Server}
\usage{
volcano_server(input, output, session, params)
}
\arguments{
\item{input}{module input}

\item{output}{module output}

\item{session}{module session}

\item{params}{parameters object with `data` and `settings` options.}
}
\value{
returns shiny module Server function
}
\description{
Modularized server for AE volcano plot.
}
