% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_ui.R
\name{volcano_ui}
\alias{volcano_ui}
\title{Volcano Plot Module - UI}
\usage{
volcano_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
returns shiny module UI
}
\description{
Modularized user interface for AE Volcano plot
}
