% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoApp.R
\name{VolcanoApp}
\alias{VolcanoApp}
\title{Volcano App}
\usage{
VolcanoApp(
  dfAE = safetyData::adam_adae,
  dfDemog = safetyData::adam_adsl,
  settings = NULL,
  runNow = TRUE
)
}
\arguments{
\item{dfAE}{AE Data}

\item{dfDemog}{demog data}

\item{settings}{safetyGraphics settings}
}
\description{
Volcano App
}
