% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlotEnh}
\alias{volcanoPlotEnh}
\title{Create a volcano plot with EnhancedVolcano}
\usage{
volcanoPlotEnh(data)
}
\arguments{
\item{data}{A data frame from getStats()}
}
\value{
a volcano plot created with ggplot
}
\description{
Create a volcano plot with EnhancedVolcano
}
\examples{
settings<-list(
  stratification_col="AEBODSYS",
  group_col="ARM",
  reference_group="Placebo",
  comparison_group="Xanomeline High Dose",
  id_col="USUBJID"
)
stats<-getStats(dfAE=safetyData::adam_adae, dfDemog = safetyData::adam_adsl, settings)
volcanoPlot(stats)

}
