% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_treatment_groups.R
\name{add_treatment_groups}
\alias{add_treatment_groups}
\title{Customize [ dm ] mapping for stats calculation.}
\usage{
add_treatment_groups(mapping, data, stratification_col)
}
\arguments{
\item{mapping}{`list` Domain metadata}

\item{data}{`data.frame` Domain data}

\item{stratification_col}{`character` Stratification column name}
}
\description{
Customize [ dm ] mapping for stats calculation.
}
