% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoApp.R
\name{volcanoApp}
\alias{volcanoApp}
\title{Volcano App}
\usage{
volcanoApp(
  dfAE = safetyData::adam_adae,
  dfDemog = safetyData::adam_adsl,
  settings = NULL,
  runNow = TRUE
)
}
\arguments{
\item{dfAE}{AE Data}

\item{dfDemog}{demog data}

\item{settings}{safetyGraphics settings}

\item{runNow}{run app immediately?}
}
\value{
Initializes Shiny app. No return value.
}
\description{
Initializes stand-alone volcano plot shiny application.
}
