% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Create a volcano plot}
\usage{
volcanoPlot(data, plotly = TRUE)
}
\arguments{
\item{data}{A data frame from getStats()}

\item{plotly}{wrap output in ggplotly? default=TRUE}
}
\value{
a volcano plot created with ggplot or plotly
}
\description{
Create a volcano plot
}
\examples{
settings<-list(
  stratification_col="AEBODSYS",
  group_col="ARM", 
  reference_group="Placebo",
  comparison_group="Xanomeline High Dose",
  id_col="USUBJID"
)
data<-read.csv('./data/test_data.csv')
stats<-getStats(dfAE=safetyData::adam_ae, dfDemog = safetyData::adam_dm, mapping)
volcanoPlot(stats)

}
